Refactor to feature-based module structure
