+++ src/features/expense/types/index.ts
@@ -1,0 +1,7 @@
+export interface Expense {
+  id: string;
+  amount: number;
+  category: string;
+  note: string;
+  date: string;
+}
